requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework-jwt
drf-jwt
PyJWT
pytz
sqlparse
django_extensions
django-request-logging
python-slugify[unidecode]
drf_rw_serializers
bleach
